cerbos/requirements.txt
flask==2.2.2
black==22.10.0
pyjwt==2.6.0
cerbos==0.4.1
